backend/requirements.txt
fastapi
uvicorn
pydantic
gitpython
httpx
